+++ src/graph_ensembles/graphs.py
@@ -4,8 +4,10 @@ ensembles and the exploration of their properties. """
 import numpy as np
 from numpy.lib.recfunctions import rec_append_fields as append_fields
 import pandas as pd
-import graph_ensembles.methods as mt
+from . import methods as mt
+from . import lib
 import warnings
+import networkx as nx
 
 
 class Graph():
@@ -434,6 +436,48 @@ class DirectedGraph(sGraph):
         if get:
             return self.gv.in_degree
 
+    def adjacency_matrix(self, kind='csr'):
+        """ Return the adjacency matrix as a scipy sparse matrix.
+        """
+        return lib.to_sparse(
+            self.e, (self.num_vertices, self.num_vertices), kind=kind,
+            i_col='src', j_col='dst', data_col=np.ones(len(self.e)))
+
+    def to_networkx(self, original=False):
+        G = nx.DiGraph()
+        if original:
+            id_conv = list(self.id_dict.keys())
+
+            if hasattr(self, 'gv'):
+                group_conv = list(self.group_dict.keys())
+                v_num = mt.id_attr_dict(
+                    self.v, id_col='id', attr_cols=['group'])
+                v = []
+                for row in v_num:
+                    v.append((id_conv[row[0]],
+                             {'group': group_conv[row[1]['group']]}))
+            else:
+                v_num = self.v.id
+                v = []
+                for row in v_num:
+                    v.append(id_conv[row])
+
+            e = []
+            for row in self.e[['src', 'dst']]:
+                e.append((id_conv[row[0]], id_conv[row[1]]))
+        else:
+            if hasattr(self, 'gv'):
+                v = mt.id_attr_dict(self.v, id_col='id', attr_cols=['group'])
+            else:
+                v = self.v.id
+
+            e = self.e[['src', 'dst']]
+
+        G.add_nodes_from(v)
+        G.add_edges_from(e)
+
+        return G
+
 
 class WeightedGraph(DirectedGraph):
     """ General class for directed graphs with weighted edges.
@@ -620,6 +664,48 @@ class WeightedGraph(DirectedGraph):
 
         if get:
             return self.gv.in_strength
+
+    def adjacency_matrix(self, kind='csr'):
+        """ Return the adjacency matrix as a scipy sparse matrix.
+        """
+        return lib.to_sparse(
+            self.e, (self.num_vertices, self.num_vertices), kind=kind,
+            i_col='src', j_col='dst', data_col='weight')
+
+    def to_networkx(self, original=False):
+        G = nx.DiGraph()
+        if original:
+            id_conv = list(self.id_dict.keys())
+
+            if hasattr(self, 'gv'):
+                group_conv = list(self.group_dict.keys())
+                v_num = mt.id_attr_dict(
+                    self.v, id_col='id', attr_cols=['group'])
+                v = []
+                for row in v_num:
+                    v.append((id_conv[row[0]],
+                             {'group': group_conv[row[1]['group']]}))
+            else:
+                v_num = self.v.id
+                v = []
+                for row in v_num:
+                    v.append(id_conv[row])
+
+            e = []
+            for row in self.e[['src', 'dst', 'weight']]:
+                e.append((id_conv[row[0]], id_conv[row[1]], row[2]))
+        else:
+            if hasattr(self, 'gv'):
+                v = mt.id_attr_dict(self.v, id_col='id', attr_cols=['group'])
+            else:
+                v = self.v.id
+
+            e = self.e[['src', 'dst', 'weight']]
+
+        G.add_nodes_from(v)
+        G.add_weighted_edges_from(e)
+
+        return G
 
 
 class LabelVertexList():
@@ -812,6 +898,63 @@ class LabelGraph(DirectedGraph):
         if get:
             return self.lv.in_degree
 
+    def adjacency_matrix(self, kind='csr', compressed=False):
+        """ Return the adjacency matrices for each label as a list of scipy
+        sparse matrices.
+        """
+        if compressed:
+            e = np.unique(self.e[['src', 'dst']])
+            adj = lib.to_sparse(
+                    e, (self.num_vertices, self.num_vertices), kind=kind,
+                    i_col='src', j_col='dst', data_col=np.ones(len(e)))
+        else:
+            adj = []
+            for i in range(self.num_labels):
+                e = self.e[self.e.label == i]
+                adj.append(lib.to_sparse(
+                    e, (self.num_vertices, self.num_vertices), kind=kind,
+                    i_col='src', j_col='dst', data_col=np.ones(len(e))))
+        return adj
+
+    def to_networkx(self, original=False):
+        G = nx.MultiDiGraph()
+        if original:
+            id_conv = list(self.id_dict.keys())
+            label_conv = list(self.label_dict.keys())
+
+            if hasattr(self, 'gv'):
+                group_conv = list(self.group_dict.keys())
+                v_num = mt.id_attr_dict(
+                    self.v, id_col='id', attr_cols=['group'])
+                v = []
+                for row in v_num:
+                    v.append((id_conv[row[0]],
+                             {'group': group_conv[row[1]['group']]}))
+            else:
+                v_num = self.v.id
+                v = []
+                for row in v_num:
+                    v.append(id_conv[row])
+
+            e = []
+            for row in self.e:
+                e.append((id_conv[row.src], id_conv[row.dst],
+                         {'label': label_conv[row.label]}))
+        else:
+            if hasattr(self, 'gv'):
+                v = mt.id_attr_dict(self.v, id_col='id', attr_cols=['group'])
+            else:
+                v = self.v.id
+
+            e = []
+            for row in self.e:
+                e.append((row.src, row.dst, {'label': row.label}))
+
+        G.add_nodes_from(v)
+        G.add_edges_from(e)
+
+        return G
+
 
 class WeightedLabelGraph(WeightedGraph, LabelGraph):
     """ General class for directed graphs with labelled and weighted edges.
@@ -948,3 +1091,62 @@ class WeightedLabelGraph(WeightedGraph, LabelGraph):
 
         if get:
             return self.lv.in_strength
+
+    def adjacency_matrix(self, kind='csr', compressed=False):
+        """ Return the adjacency matrices for each label as a list of scipy
+        sparse matrices.
+        """
+        if compressed:
+            adj = lib.to_sparse(
+                    self.e, (self.num_vertices, self.num_vertices), kind=kind,
+                    i_col='src', j_col='dst', data_col='weight')
+        else:
+            adj = []
+            for i in range(self.num_labels):
+                e = self.e[self.e.label == i]
+                adj.append(lib.to_sparse(
+                    e, (self.num_vertices, self.num_vertices), kind=kind,
+                    i_col='src', j_col='dst', data_col='weight'))
+        return adj
+
+    def to_networkx(self, original=False):
+        G = nx.MultiDiGraph()
+        if original:
+            id_conv = list(self.id_dict.keys())
+            label_conv = list(self.label_dict.keys())
+
+            if hasattr(self, 'gv'):
+                group_conv = list(self.group_dict.keys())
+                v_num = mt.id_attr_dict(
+                    self.v, id_col='id', attr_cols=['group'])
+                v = []
+                for row in v_num:
+                    v.append((id_conv[row[0]],
+                             {'group': group_conv[row[1]['group']]}))
+            else:
+                v_num = self.v.id
+                v = []
+                for row in v_num:
+                    v.append(id_conv[row])
+
+            e = []
+            for row in self.e:
+                e.append((id_conv[row.src], id_conv[row.dst],
+                         {'weight': row.weight,
+                          'label': label_conv[row.label]}))
+        else:
+            if hasattr(self, 'gv'):
+                v = mt.id_attr_dict(self.v, id_col='id', attr_cols=['group'])
+            else:
+                v = self.v.id
+
+            e = []
+            for row in self.e:
+                e.append((row.src, row.dst,
+                         {'weight': row.weight,
+                          'label': row.label}))
+
+        G.add_nodes_from(v)
+        G.add_edges_from(e)
+
+        return G